% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapepnet.r
\name{getPNetPredictions}
\alias{getPNetPredictions}
\title{\code{getPNetPredictions} returns a data frame with predicted upstream kinases obtained from the phosphoNET website}
\usage{
getPNetPredictions(urlBase = "http://www.phosphonet.ca/kinasepredictor.aspx",
  uni, ps)
}
\arguments{
\item{urlBase}{url base string for phospoNET html query. Default: "http://www.phosphonet.ca/kinasepredictor.aspx"}

\item{uni}{a uniprot ID}

\item{ps}{phs]ophorylated residue}
}
\value{
a data frame with the phosphoNET prediction table. Returns NULL if the uni ps combination is not found.
}
\description{
\code{getPNetPredictions} returns a data frame with predicted upstream kinases obtained from the phosphoNET website
}
\examples{
# For obtaining predictions for Tyrosine 486 form protein with uniprot ID Q14247
getPNetPredictions(uni = "Q14247", ps = "Y486")
}

