% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapepnet.r
\name{getPNetPredictionTable}
\alias{getPNetPredictionTable}
\title{\code{getPNetPredictions} returns a data frame with predicted upstream kinases obtained from the phosphoNET website for multiple phosphosites}
\usage{

  getPNetPredictionTable(urlBase = "http://www.phosphonet.ca/kinasepredictor.aspx",
  uni, ps)
}
\arguments{
\item{urlBase}{url base string for phospoNET html query. Default: "http://www.phosphonet.ca/kinasepredictor.aspx"}

\item{uni}{a vectyor containing uniprot IDs}

\item{ps}{a vector containing phosphorylated residues, same length as uni}
}
\value{
a data frame with the phosphoNET predictions for the uni ps combination that were found found.
}
\description{
\code{getPNetPredictions} returns a data frame with predicted upstream kinases obtained from the phosphoNET website for multiple phosphosites
}
\examples{
# For obtaining predictions for Tyrosine 486, Threonine 24, ans Serine 11 of protein with uniprot ID Q14247
getPNetPredictionTable(uni = c("Q14247", "Q14247","Q14247"), ps = c("Y486", "T24","S11"))
}

